+++ frontend/src/app/features/main-page/components/description-section/components/game/game.component.ts
@@ -1,8 +1,11 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { GameCategory } from '../../../../../../interfaces/game-category.interface';
+import { EditGameService } from '../../../../services/edit-game.service';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 
 @Component({
     selector: 'mados-game',
     templateUrl: './game.component.html',
+    styleUrls: ['./game.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class GameComponent {
@@ -11,5 +14,13 @@ export class GameComponent {
     @Input()
     public name: string;
     @Input()
-    public iconNames: string[] | null;
+    public categories: GameCategory[] | undefined;
+    @Input()
+    public shouldShowEdition: boolean;
+
+    constructor(private editGameService: EditGameService) { }
+
+    public editGame(): void {
+        this.editGameService.editingGameName$.next(this.name);
+    }
 }