+++ frontend/src/app/interfaces/game.interface.ts
@@ -1,5 +1,8 @@
+import { CategoryType } from '../enums/category-type.enum';
+import { GameCategory } from './game-category.interface';
+
 export interface Game {
     oneHundredPercentIconName: string;
     name: string;
-    iconNames: string[] | null;
+    categories?: GameCategory[];
 }